Add tm script for opening project in tmux and Vim
While watching one of his videos, I noticed ThePrimeagen firing up new
tmux sessions via fzf, which appeared to have been launched with a list
of the folders that he might want to open. I suspect he was using [1].

tm is my take on this idea.

I know that all the folders I want to open in Vim live inside my
`~/Projects` folder, and they'll all have a `.git` directory within
them. So that's the set of folders that I pass to fzf.

It will either attach to an existing tmux session that's named after the
last part of the directory name, or launch Vim in a new session.

I've written it so others can customise it to their directory layout, or
for different version control systems. Set the TM_ environment variables
accordingly.

[1] https://github.com/ThePrimeagen/.dotfiles/blob/master/bin/.local/bin/tmux-sessionizer