tm can be run from within tmux
You can use the `tmux` command to switch to a running session from
within another tmux session, using the `switch-client` command.

If you're not yet within a session you can attach to a running session
with the `attach-session` command.

This means that we can "attach" with one of those two commands, once
we've ensured that the session exists.